+++ lib/danger/request_sources/bitbucket_cloud_api.rb
@@ -5,13 +5,13 @@ require "danger/helpers/comments_helper"
 module Danger
   module RequestSources
     class BitbucketCloudAPI
-      attr_accessor :host, :pr_api_endpoint, :pr_api_endpoint_v1
+      attr_accessor :project, :slug, :pull_request_id
 
-      def initialize(project, slug, pull_request_id, environment)
+      def initialize(repo_slug, pull_request_id, branch_name, environment)
         @username = environment["DANGER_BITBUCKETCLOUD_USERNAME"]
         @password = environment["DANGER_BITBUCKETCLOUD_PASSWORD"]
-        self.pr_api_endpoint = "https://api.bitbucket.org/2.0/repositories/#{project}/#{slug}/pullrequests/#{pull_request_id}"
-        self.pr_api_endpoint_v1 = "https://api.bitbucket.org/1.0/repositories/#{project}/#{slug}/pullrequests/#{pull_request_id}"
+        self.project, self.slug = repo_slug.split("/")
+        self.pull_request_id = pull_request_id || fetch_pr_from_branch(branch_name)
       end
 
       def inspect
@@ -50,6 +50,27 @@ module Danger
       end
 
       private
+
+      def base_url(version)
+        "https://api.bitbucket.org/#{version}.0/repositories/#{project}/#{slug}/pullrequests"
+      end
+
+      def pr_api_endpoint
+        "#{base_url(2)}/#{pull_request_id}"
+      end
+
+      def pr_api_endpoint_v1
+        "#{base_url(1)}/#{pull_request_id}"
+      end
+
+      def prs_api_endpoint(branch_name)
+        "#{base_url(2)}?q=source.branch.name=\"#{branch_name}\""
+      end
+
+      def fetch_pr_from_branch(branch_name)
+        uri = URI(URI.escape(prs_api_endpoint(branch_name)))
+        fetch_json(uri)[:values][0][:id]
+      end
 
       def fetch_json(uri)
         req = Net::HTTP::Get.new(uri.request_uri, { "Content-Type" => "application/json" })